adds in a new pair of female underwear
The sprite is an old sprite that got removed in this commit:
https://github.com/tgstation/tgstation/commit/ca41b444f36659016137bc18a851acebd4318b64

I made some edits to it to make it fit on the female sprite better.

The original sprite seems to date all the way back to the R4407 source release, so one of Goon's spriters at the time likely made it, although i'm not sure which one.