removes the grill
The grill is utterly useless for real cooking, and it serves zero purpose beyond "hehehe le funny boomer joke XD".

It probably has a reason to exist on servers such as Fulp, or /tg/'s Terry or Basil, but it doesn't fit in on a server with higher roleplay standards.